WIP - implementing acceleration raw readings conversion routine.
