+++ drivers/kionix/kx132-1211/README-KX132-driver-roadmap.rst
@@ -15,8 +15,8 @@ kx132-1211.h          . . . primary driver header file, defines:
 
                             *  Zephyr RTOS sensor data and config structures
                             *  extra Zephyr sensor channels to read KX132 specific parameters
-                            *  custom enum to support Zephyr's sensor attribute set and get API convention
-                            *  KX132 possible I2C device addresses, per TN027-Power-On-Procedure.pdf
+                            *  sensor channel and sensor attribute enumerations to support Zephyr API conventions
+                            *  KX132 available I2C device addresses, per TN027-Power-On-Procedure.pdf
 
 kx132-1211.c          . . . implements Zephyr sensor APIS
 
@@ -47,11 +47,28 @@ kx132-registers.c     . . . implements:
                             *  implements defined sensor configuration sequences
                             *  implements sensor parameter fetch (read) routines
 
-kx132-triggers.[ch]   . . . 
+kx132-triggers.[ch]   . . . NOT IMPLEMENTED NOT TESTED TO WORK, AS OF 2023 Q1 (note 1) - TMH
 
-out-of-tree-drivers.h
-                      . . . enumerated driver routine return values
+out-of-tree-drivers.h . . . enumerated driver routine return values
 
+
+REFERENCES
+**********
+
+*  TN027-Power-On-Procedure.pdf                        . . . recommended power up test for all uses, I2C addressing details
+*  KX132-1211-Technical-Reference-Manual-Rev-5.0.pdf   . . . register and sensor function details
+*  AN092-Getting-Started.pdf                           . . . application notes on KX132 practical use configurations
+
+*  https://github.com/zephyrproject-rtos/zephyr/tree/main/samples/sensor/lis2dh
+*  https://github.com/zephyrproject-rtos/zephyr/blob/main/samples/sensor/lis2dh/src/main.c#L81
+
+*  https://github.com/zephyrproject-rtos/zephyr/tree/main/samples/basic/button
+
+
+NOTES
+*****
+
+(1)  Developer attempts to configure a GPIO based interrupt for KX132 driver, within Zephyr's sensor driver framework and practices have failed as of 2022 Q4.  In other words, attempts to configure a hardware interrupt in this KX132 driver, following example of Zephyr 3.2.0 sample app for STMicro's LIS2DH sensor have so far not succeeded.  An alternative however is tested and working, to set up in one's Zephyr app a GPIO based interrupt in the application.  An example of code for this sensor interrupt solutino is exemplified in Zephyr 3.2.0 sample application named 'button'.
 
 
 CONTRIBUTORS
