+++ drivers/kionix/kx132-1211/kx132-registers.h
@@ -267,6 +267,16 @@ int kx132_get_attr__buf_cntl1__sample_threshold_setting(const struct device *dev
 
 int kx132_get_attr__buf_read__sample_as_attribute(const struct device *dev, struct sensor_value *val);
 
+// Not technically a attribute getters, but returns which return a value immediately:
+
+/**
+ * @brief Routine which takes a raw accelerometer reading, and full
+ *        range value, e.g. +2.0g, +4.0g, etc, and returns an
+ *        acceleration value in units of g.
+ */
+int kx132_get_attr__acc_reading_in_g(const struct device *dev, struct sensor_value *val);
+
+
 
 // - GROUP - routines which write values to sensor registers:
